refactor : config 폴더정리하면서 바뀐 경로 적용
