+++ src/main/java/com/web/SearchWeb/bookmark/service/BookmarkServiceImpl.java
@@ -15,8 +15,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.web.SearchWeb.bookmark.error.BookmarkErrorCode;
-import com.web.SearchWeb.config.BusinessException;
-import com.web.SearchWeb.config.CommonErrorCode;
+import com.web.SearchWeb.config.exception.BusinessException;
+import com.web.SearchWeb.config.exception.CommonErrorCode;
 
 import com.web.SearchWeb.bookmark.dto.MemberTagResultDto;
 import java.net.URI;
