+++ src/main/java/com/web/SearchWeb/tag/error/TagException.java
@@ -1,7 +1,7 @@
 package com.web.SearchWeb.tag.error;
 
-import com.web.SearchWeb.config.BusinessException;
-import com.web.SearchWeb.config.ErrorCode;
+import com.web.SearchWeb.config.exception.BusinessException;
+import com.web.SearchWeb.config.exception.ErrorCode;
 import lombok.Getter;
 
 @Getter
